refactor(admin): split user resource panel
